structure win32 crate
