+++ projs/handmade_rust/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "handmade_rust"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+win32 = {path= "../win32"}
