+++ projs/win32/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "win32"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
