+++ projs/win32/src/helpers.rs
@@ -1,0 +1,4 @@
+/// Turns a Rust string slice into a null-terminated utf-16 vector.
+pub fn wide_null(s: &str) -> Vec<u16> {
+    s.encode_utf16().chain(Some(0)).collect()
+}
